+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.1.0
+- The Xiangqi Update!
+- `BoardConfig` - used to configure the way `BoardBackground` is drawn.
+- Two Xiangqi piece sets and one background image.
+
 ### 1.0.5
 - Fixed move string parsing for moves from ranks larger than 9 (thanks @malaschitz).
 - Fixed duplicate pieces appearing in piece selector for castling gating moves.